+++ isik/django/drf/__init__.py
@@ -7,11 +7,14 @@ from isik.django.drf.error_handling import django_to_drf_validation_error  # noq
 from isik.django.drf.filters import make_filters  # noqa: E402
 from isik.django.drf.pagination import PageNumberPagination  # noqa: E402
 from isik.django.drf.permissions import (  # noqa: E402
+    Guard,
     IsAnonymous,
     IsAuthenticatedANDSignupCompleted,
     IsSuperUser,
     ReadOnly,
+    guarding,
     is_owner,
+    object_property,
     prevent_actions,
     user_property,
 )
@@ -21,14 +24,17 @@ from isik.django.drf.schema import FakeErrorSerializer, FakeSerializer  # noqa: 
 __all__ = [
     "FakeErrorSerializer",
     "FakeSerializer",
+    "Guard",
     "IsAnonymous",
     "IsAuthenticatedANDSignupCompleted",
     "IsSuperUser",
     "PageNumberPagination",
     "ReadOnly",
     "django_to_drf_validation_error",
+    "guarding",
     "is_owner",
     "make_filters",
+    "object_property",
     "prevent_actions",
     "user_property",
 ]