+++ isik/django/drf/viewsets/__init__.py
@@ -1,6 +1,7 @@
 from isik.django.drf.viewsets.action_serializer_class import ActionSerializerClassMixin
 from isik.django.drf.viewsets.base import BaseModelViewSet
 from isik.django.drf.viewsets.filterset import FilterSetMixin
+from isik.django.drf.viewsets.guarded_fields import GuardedFieldsMixin
 from isik.django.drf.viewsets.history import HistoryMixin, context_filter
 from isik.django.drf.viewsets.ordering import DeclaredOrderingFilter, DeclaredOrderingMixin, ReverseOrderingMixin
 from isik.django.drf.viewsets.protected_destroy import ProtectedDestroyMixin
@@ -14,6 +15,7 @@ __all__ = [
     "DeclaredOrderingFilter",
     "DeclaredOrderingMixin",
     "FilterSetMixin",
+    "GuardedFieldsMixin",
     "HistoryMixin",
     "ProtectedDestroyMixin",
     "ReverseOrderingMixin",